rename reports/index to reports/reports
